feat: add users show endpoint

+++ routes/api/users.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\UserRoleController;
 use App\Http\Controllers\UserStatusController;
 
 Route::get('/', [UserController::class, 'index'])->name('index');
+Route::get('/{user}', [UserController::class, 'show'])->name('show');
 Route::patch('/{user}', [UserController::class, 'update'])->name('update');
 Route::delete('/{user}', [UserController::class, 'destroy'])->name('delete');
 
@@ -16,8 +17,5 @@ Route::patch('/email/update', UserEmailController::class)->name('email:update');
 Route::patch('/{user}/status', UserStatusController::class)->name('status:update');
 Route::patch('/{user}/role', UserRoleController::class)->name('role:update');
 
-Route::controller(AdminController::class)->group(function () {
-
-    Route::post('/admin', 'store')->name('admins:store');
-    Route::delete('/admin/{user}', 'destroy')->name('admins:delete');
-});
+Route::post('/admin', [AdminController::class, 'store'])->name('admins:store');
+Route::delete('/admin/{user}', [AdminController::class, 'destroy'])->name('admins:delete');
